Extract unzipping and reading of csv into a page object
